refactor(gateway): leaf observer contract and ask module split for ci gates
