+++ src/gateway/session-observer-model.ts
@@ -30,8 +30,6 @@ const MAX_DORMANT_RUNS = 256;
 const MAX_DISABLED_RUNS = 512;
 
 export const SESSION_OBSERVER_MODEL_MAX_TOKENS = 300;
-export const SESSION_OBSERVER_ASK_MODEL_MAX_TOKENS = 400;
-export const SESSION_OBSERVER_ASK_ANSWER_MAX_CHARS = 600;
 type PrepareModel = typeof prepareSimpleCompletionModelForAgent;
 type CompleteModel = typeof completeWithPreparedSimpleCompletionModel;
 export type PreparedModel = Awaited<ReturnType<PrepareModel>>;
@@ -231,12 +229,6 @@ export const SESSION_OBSERVER_SYSTEM_PROMPT = [
   'Return strict JSON only, for example: {"headline":"Checking the fix","assessment":"Tests are passing.","health":"on-track","planProgress":{"completed":2,"total":3}}. Omit optional fields instead of setting them to null.',
 ].join(" ");
 
-export const SESSION_OBSERVER_ASK_SYSTEM_PROMPT = [
-  "You answer operator questions about a running AI agent session using only the supplied observation digest and notes.",
-  "Do not infer details that are absent from the observations; plainly say when you cannot know.",
-  "Return only a concise plain-text answer in American English, with no markdown or JSON wrapper.",
-].join(" ");
-
 const ModelDigestSchema = z
   .strictObject({
     headline: z.string().min(1),
@@ -428,18 +420,6 @@ export function buildSessionObserverPrompt(
     previousDigest: state.previousDigest ?? null,
     newNotes: notes,
     planProgress: state.planProgress ?? null,
-  });
-}
-
-export function buildSessionObserverAskPrompt(params: {
-  digest?: SessionObserverDigest;
-  notes: readonly string[];
-  question: string;
-}): string {
-  return JSON.stringify({
-    digest: params.digest ?? null,
-    notes: params.notes,
-    question: params.question,
   });
 }
